src/requirements.txt
streamlit
crewai
langchain
pymupdf
python-dotenv
langchain-groq
langchain-google-genai
gradio>=4.0.0
transformers>=4.35.0